Move evaluation types back into Types. :-(
